requirements.txt
asgiref==3.8.1
async-timeout==5.0.1
attrs==25.3.0
Django==5.2
django-cors-headers==4.7.0
django-jazzmin==3.0.1
django-modeltranslation==0.19.14
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
drf-spectacular==0.28.0
gunicorn==23.0.0
inflection==0.5.1
jsonschema==4.23.0
jsonschema-specifications==2025.4.1
packaging==25.0
pillow==11.2.1
psycopg2-binary==2.9.10
PyJWT==2.9.0
python-decouple==3.8
PyYAML==6.0.2
redis==6.0.0
referencing==0.36.2
rpds-py==0.24.0
sqlparse==0.5.3
typing_extensions==4.13.2
uritemplate==4.2.0

twilio~=9.10.4